+++ x/marker/keeper/msg_server.go
@@ -623,6 +623,7 @@ func (k msgServer) AddFinalizeActivateMarker(goCtx context.Context, msg *types.M
 	return &types.MsgAddFinalizeActivateMarkerResponse{}, nil
 }
 
+// SupplyIncreaseProposal can only be called via gov proposal
 func (k msgServer) SupplyIncreaseProposal(goCtx context.Context, msg *types.MsgSupplyIncreaseProposalRequest) (*types.MsgSupplyIncreaseProposalResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
@@ -643,6 +644,7 @@ func (k msgServer) SupplyIncreaseProposal(goCtx context.Context, msg *types.MsgS
 	return &types.MsgSupplyIncreaseProposalResponse{}, nil
 }
 
+// UpdateRequiredAttributes will only succeed if signer has transfer authority
 func (k msgServer) UpdateRequiredAttributes(goCtx context.Context, msg *types.MsgUpdateRequiredAttributesRequest) (*types.MsgUpdateRequiredAttributesResponse, error) {
 	ctx := sdk.UnwrapSDKContext(goCtx)
 
@@ -699,6 +701,7 @@ func (k msgServer) UpdateRequiredAttributes(goCtx context.Context, msg *types.Ms
 	return &types.MsgUpdateRequiredAttributesResponse{}, nil
 }
 
+// UpdateForcedTransfer updates the allow_forced_transfer field of a marker via governance proposal.
 func (k msgServer) UpdateForcedTransfer(goCtx context.Context, msg *types.MsgUpdateForcedTransferRequest) (*types.MsgUpdateForcedTransferResponse, error) {
 	if msg.Authority != k.GetAuthority() {
 		return nil, errors.Wrapf(govtypes.ErrInvalidSigner, "expected %s got %s", k.GetAuthority(), msg.Authority)
@@ -733,4 +736,31 @@ func (k msgServer) UpdateForcedTransfer(goCtx context.Context, msg *types.MsgUpd
 	)
 
 	return &types.MsgUpdateForcedTransferResponse{}, nil
+}
+
+// SetAccountData sets the accountdata for a denom. Signer must have deposit authority.
+func (k msgServer) SetAccountData(goCtx context.Context, msg *types.MsgSetAccountDataRequest) (*types.MsgSetAccountDataResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	marker, err := k.GetMarkerByDenom(ctx, msg.Denom)
+	if err != nil {
+		return nil, fmt.Errorf("could not get %s marker: %w", msg.Denom, err)
+	}
+
+	if msg.Signer == k.GetAuthority() {
+		if !marker.HasGovernanceEnabled() {
+			return nil, fmt.Errorf("%s marker does not allow governance control", msg.Denom)
+		}
+	} else {
+		if !marker.HasAccess(msg.Signer, types.Access_Deposit) {
+			return nil, fmt.Errorf("%s does not have deposit access for %s marker", msg.Signer, msg.Denom)
+		}
+	}
+
+	err = k.attrKeeper.SetAccountData(ctx, marker.GetAddress().String(), msg.Value)
+	if err != nil {
+		return nil, fmt.Errorf("error setting %s account data: %w", msg.Denom, err)
+	}
+
+	return &types.MsgSetAccountDataResponse{}, nil
 }