+++ x/marker/simulation/operations_test.go
@@ -1,7 +1,9 @@
 package simulation_test
 
 import (
+	"bytes"
 	"math/rand"
+	"strings"
 	"testing"
 	"time"
 
@@ -14,6 +16,7 @@ import (
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	"github.com/cosmos/cosmos-sdk/x/bank/testutil"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
+	"github.com/provenance-io/provenance/x/marker/keeper"
 
 	abci "github.com/tendermint/tendermint/abci/types"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
@@ -31,23 +34,41 @@ type SimTestSuite struct {
 	app *app.App
 }
 
-func (suite *SimTestSuite) SetupTest() {
-	suite.app = app.Setup(suite.T())
-	suite.ctx = suite.app.BaseApp.NewContext(false, tmproto.Header{})
+func (s *SimTestSuite) SetupTest() {
+	s.app = app.Setup(s.T())
+	s.ctx = s.app.BaseApp.NewContext(false, tmproto.Header{})
 }
 
-func (suite *SimTestSuite) TestWeightedOperations() {
-	cdc := suite.app.AppCodec()
+// LogOperationMsg logs all fields of the provided operationMsg.
+func (s *SimTestSuite) LogOperationMsg(operationMsg simtypes.OperationMsg) {
+	msgFmt := "%s"
+	if len(bytes.TrimSpace(operationMsg.Msg)) == 0 {
+		msgFmt = "    %q"
+	}
+	fmtLines := []string{
+		"operationMsg.Route:   %q",
+		"operationMsg.Name:    %q",
+		"operationMsg.Comment: %q",
+		"operationMsg.OK:      %t",
+		"operationMsg.Msg: " + msgFmt,
+	}
+	s.T().Logf(strings.Join(fmtLines, "\n"),
+		operationMsg.Route, operationMsg.Name, operationMsg.Comment, operationMsg.OK, string(operationMsg.Msg),
+	)
+}
+
+func (s *SimTestSuite) TestWeightedOperations() {
+	cdc := s.app.AppCodec()
 	appParams := make(simtypes.AppParams)
 
-	weightedOps := simulation.WeightedOperations(appParams, cdc, codec.NewProtoCodec(suite.app.InterfaceRegistry()), suite.app.MarkerKeeper,
-		suite.app.AccountKeeper, suite.app.BankKeeper, suite.app.GovKeeper,
+	weightedOps := simulation.WeightedOperations(appParams, cdc, codec.NewProtoCodec(s.app.InterfaceRegistry()), s.app.MarkerKeeper,
+		s.app.AccountKeeper, s.app.BankKeeper, s.app.GovKeeper, s.app.AttributeKeeper,
 	)
 
 	// setup 3 accounts
-	s := rand.NewSource(1)
-	r := rand.New(s)
-	accs := suite.getTestingAccounts(r, 3)
+	src := rand.NewSource(1)
+	r := rand.New(src)
+	accs := s.getTestingAccounts(r, 3)
 
 	// Note: r is now passed around and used in several places, including in SDK functions.
 	//	Since we're seeding it, these tests are deterministic. However, if there are changes
@@ -77,94 +98,86 @@ func (suite *SimTestSuite) TestWeightedOperations() {
 		{simappparams.DefaultWeightMsgAddAccess, sdk.MsgTypeURL(&types.MsgAddAccessRequest{}), sdk.MsgTypeURL(&types.MsgAddAccessRequest{})},
 		{simappparams.DefaultWeightMsgAddFinalizeActivateMarker, sdk.MsgTypeURL(&types.MsgAddFinalizeActivateMarkerRequest{}), sdk.MsgTypeURL(&types.MsgAddFinalizeActivateMarkerRequest{})},
 		{simappparams.DefaultWeightMsgAddMarkerProposal, "gov", sdk.MsgTypeURL(&govtypes.MsgSubmitProposal{})},
+		{simappparams.DefaultWeightMsgSetAccountData, sdk.MsgTypeURL(&types.MsgSetAccountDataRequest{}), sdk.MsgTypeURL(&types.MsgSetAccountDataRequest{})},
 	}
+
+	expNames := make([]string, len(expected))
+	for i, exp := range expected {
+		expNames[i] = exp.opMsgName
+	}
+
+	// Run all the ops and get the operation messages and their names.
+	opMsgs := make([]simtypes.OperationMsg, len(weightedOps))
+	actualNames := make([]string, len(weightedOps))
 	for i, w := range weightedOps {
-		operationMsg, _, _ := w.Op()(r, suite.app.BaseApp, suite.ctx, accs, "")
-		// the following checks are very much dependent from the ordering of the output given
-		// by WeightedOperations. if the ordering in WeightedOperations changes some tests
-		// will fail
-		suite.Require().Equal(expected[i].weight, w.Weight(), "weight should be the same")
-		suite.Require().Equal(expected[i].opMsgRoute, operationMsg.Route, "route should be the same")
-		suite.Require().Equal(expected[i].opMsgName, operationMsg.Name, "operation Msg name should be the same")
+		opMsgs[i], _, _ = w.Op()(r, s.app.BaseApp, s.ctx, accs, "")
+		actualNames[i] = opMsgs[i].Name
+	}
+
+	// First, make sure the op names are as expected since a failure there probably means the rest will fail.
+	// And it's probably easier to address when you've got a nice list comparison of names and their orderings.
+	s.Require().Equal(expNames, actualNames, "operation message names")
+
+	for i := range expected {
+		s.Require().Equal(expected[i].weight, weightedOps[i].Weight(), "weightedOps[i].Weight", i)
+		s.Require().Equal(expected[i].opMsgRoute, opMsgs[i].Route, "opMsgs[i].Route", i)
+		s.Require().Equal(expected[i].opMsgName, opMsgs[i].Name, "opMsgs[i].Name", i)
 	}
 }
 
 // TestSimulateMsgAddMarker tests the normal scenario of a valid message of type TypeAddMarkerRequest.
 // Abnormal scenarios, where the message is created by an errors, are not tested here.
-func (suite *SimTestSuite) TestSimulateMsgAddMarker() {
-
+func (s *SimTestSuite) TestSimulateMsgAddMarker() {
 	// setup 3 accounts
-	s := rand.NewSource(1)
-	r := rand.New(s)
-	accounts := suite.getTestingAccounts(r, 3)
+	src := rand.NewSource(1)
+	r := rand.New(src)
+	accounts := s.getTestingAccounts(r, 3)
 
 	// begin a new block
-	suite.app.BeginBlock(abci.RequestBeginBlock{Header: tmproto.Header{Height: suite.app.LastBlockHeight() + 1, AppHash: suite.app.LastCommitID().Hash}})
+	s.app.BeginBlock(abci.RequestBeginBlock{Header: tmproto.Header{Height: s.app.LastBlockHeight() + 1, AppHash: s.app.LastCommitID().Hash}})
 
 	// execute operation
-	op := simulation.SimulateMsgAddMarker(suite.app.MarkerKeeper, suite.app.AccountKeeper, suite.app.BankKeeper)
-	operationMsg, futureOperations, err := op(r, suite.app.BaseApp, suite.ctx, accounts, "")
-	suite.Require().NoError(err)
+	op := simulation.SimulateMsgAddMarker(s.app.MarkerKeeper, s.app.AccountKeeper, s.app.BankKeeper)
+	operationMsg, futureOperations, err := op(r, s.app.BaseApp, s.ctx, accounts, "")
+	s.Require().NoError(err)
 
 	var msg types.MsgAddMarkerRequest
 	types.ModuleCdc.UnmarshalJSON(operationMsg.Msg, &msg)
 
-	suite.Require().True(operationMsg.OK, operationMsg.String())
-	suite.Require().Equal(sdk.MsgTypeURL(&msg), operationMsg.Name)
-	suite.Require().Equal(sdk.MsgTypeURL(&msg), operationMsg.Route)
-	suite.Require().Len(futureOperations, 0)
+	s.Require().True(operationMsg.OK, operationMsg.String())
+	s.Require().Equal(sdk.MsgTypeURL(&msg), operationMsg.Name)
+	s.Require().Equal(sdk.MsgTypeURL(&msg), operationMsg.Route)
+	s.Require().Len(futureOperations, 0)
 }
 
 // TestSimulateMsgAddActivateFinalizeMarker tests the normal scenario of a valid message of type TypeAddActivateFinalizeMarkerRequest.
 // Abnormal scenarios, where the message is created by an errors, are not tested here.
-func (suite *SimTestSuite) TestSimulateMsgAddActivateFinalizeMarker() {
-
+func (s *SimTestSuite) TestSimulateMsgAddActivateFinalizeMarker() {
 	// setup 3 accounts
-	s := rand.NewSource(1)
-	r := rand.New(s)
-	accounts := suite.getTestingAccounts(r, 3)
+	src := rand.NewSource(1)
+	r := rand.New(src)
+	accounts := s.getTestingAccounts(r, 3)
 
 	// begin a new block
-	suite.app.BeginBlock(abci.RequestBeginBlock{Header: tmproto.Header{Height: suite.app.LastBlockHeight() + 1, AppHash: suite.app.LastCommitID().Hash}})
+	s.app.BeginBlock(abci.RequestBeginBlock{Header: tmproto.Header{Height: s.app.LastBlockHeight() + 1, AppHash: s.app.LastCommitID().Hash}})
 
 	// execute operation
-	op := simulation.SimulateMsgAddFinalizeActivateMarker(suite.app.MarkerKeeper, suite.app.AccountKeeper, suite.app.BankKeeper)
-	operationMsg, futureOperations, err := op(r, suite.app.BaseApp, suite.ctx, accounts, "")
-	suite.Require().NoError(err)
+	op := simulation.SimulateMsgAddFinalizeActivateMarker(s.app.MarkerKeeper, s.app.AccountKeeper, s.app.BankKeeper)
+	operationMsg, futureOperations, err := op(r, s.app.BaseApp, s.ctx, accounts, "")
+	s.Require().NoError(err)
 
 	var msg types.MsgAddFinalizeActivateMarkerRequest
 	types.ModuleCdc.UnmarshalJSON(operationMsg.Msg, &msg)
 
-	suite.Require().True(operationMsg.OK, operationMsg.String())
-	suite.Require().Equal(sdk.MsgTypeURL(&msg), operationMsg.Name)
-	suite.Require().Equal(sdk.MsgTypeURL(&msg), operationMsg.Route)
-	suite.Require().Len(futureOperations, 0)
-}
-
-func (suite *SimTestSuite) getTestingAccounts(r *rand.Rand, n int) []simtypes.Account {
-	accounts := simtypes.RandomAccounts(r, n)
-
-	initAmt := sdk.TokensFromConsensusPower(1000000, sdk.DefaultPowerReduction)
-	initCoins := sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, initAmt))
-
-	// add coins to the accounts
-	for _, account := range accounts {
-		acc := suite.app.AccountKeeper.NewAccountWithAddress(suite.ctx, account.Address)
-		suite.app.AccountKeeper.SetAccount(suite.ctx, acc)
-		err := testutil.FundAccount(suite.app.BankKeeper, suite.ctx, account.Address, initCoins)
-		suite.Require().NoError(err)
-	}
-
-	return accounts
-}
-
-func TestSimTestSuite(t *testing.T) {
-	suite.Run(t, new(SimTestSuite))
+	s.Require().True(operationMsg.OK, operationMsg.String())
+	s.Require().Equal(sdk.MsgTypeURL(&msg), operationMsg.Name)
+	s.Require().Equal(sdk.MsgTypeURL(&msg), operationMsg.Route)
+	s.Require().Len(futureOperations, 0)
 }
 
 // TestSimulateMsgAddMarkerProposal tests the normal scenario of a valid message of type MsgAddMarkerProposalRequest.
 // Abnormal scenarios, where the message is created by an errors, are not tested here.
-func (suite *SimTestSuite) TestSimulateMsgAddMarkerProposal() {
+func (s *SimTestSuite) TestSimulateMsgAddMarkerProposal() {
 	NewMsgAddMarker := func(denom string, totalSupply sdkmath.Int, manager sdk.AccAddress, status types.MarkerStatus,
 		markerType types.MarkerType, access []types.AccessGrant, fixed bool, allowGov bool, authority string,
 	) *types.MsgAddMarkerRequest {
@@ -183,14 +196,14 @@ func (suite *SimTestSuite) TestSimulateMsgAddMarkerProposal() {
 		}
 	}
 
-	s := rand.NewSource(1)
-	r := rand.New(s)
-	accounts := suite.getTestingAccounts(r, 10)
-	accounts = append(accounts, suite.createTestingAccountsWithPower(r, 1, 0)...)
-	accounts = append(accounts, suite.createTestingAccountsWithPower(r, 1, 1)...)
+	src := rand.NewSource(1)
+	r := rand.New(src)
+	accounts := s.getTestingAccounts(r, 10)
+	accounts = append(accounts, s.createTestingAccountsWithPower(r, 1, 0)...)
+	accounts = append(accounts, s.createTestingAccountsWithPower(r, 1, 1)...)
 	acctZero := accounts[len(accounts)-2]
 	acctOne := accounts[len(accounts)-1]
-	acctOneBalance := suite.app.BankKeeper.SpendableCoins(suite.ctx, acctOne.Address)
+	acctOneBalance := s.app.BankKeeper.SpendableCoins(s.ctx, acctOne.Address)
 	var acctOneBalancePlusOne sdk.Coins
 	for _, c := range acctOneBalance {
 		acctOneBalancePlusOne = acctOneBalancePlusOne.Add(sdk.NewCoin(c.Denom, c.Amount.AddRaw(1)))
@@ -201,8 +214,8 @@ func (suite *SimTestSuite) TestSimulateMsgAddMarkerProposal() {
 	depositPeriod := 1 * time.Second
 
 	resetParams := func(t *testing.T, ctx sdk.Context) {
-		require.NotPanics(suite.T(), func() {
-			suite.app.GovKeeper.SetDepositParams(suite.ctx, govtypes.DepositParams{
+		require.NotPanics(s.T(), func() {
+			s.app.GovKeeper.SetDepositParams(s.ctx, govtypes.DepositParams{
 				MinDeposit:       govMinDep,
 				MaxDepositPeriod: &depositPeriod,
 			})
@@ -294,14 +307,14 @@ func (suite *SimTestSuite) TestSimulateMsgAddMarkerProposal() {
 	}
 
 	for _, tc := range tests {
-		resetParams(suite.T(), suite.ctx)
-		suite.Run(tc.name, func() {
+		resetParams(s.T(), s.ctx)
+		s.Run(tc.name, func() {
 
 			args := &simulation.SendGovMsgArgs{
-				WeightedOpsArgs: suite.getWeightedOpsArgs(),
+				WeightedOpsArgs: s.getWeightedOpsArgs(),
 				R:               rand.New(rand.NewSource(1)),
-				App:             suite.app.BaseApp,
-				Ctx:             suite.ctx,
+				App:             s.app.BaseApp,
+				Ctx:             s.ctx,
 				Accs:            accounts,
 				ChainID:         "send-gov-test",
 				Sender:          tc.sender,
@@ -316,21 +329,21 @@ func (suite *SimTestSuite) TestSimulateMsgAddMarkerProposal() {
 			testFunc := func() {
 				skip, opMsg, _ = simulation.SendGovMsg(args)
 			}
-			suite.Require().NotPanics(testFunc, "SendGovMsg")
+			s.Require().NotPanics(testFunc, "SendGovMsg")
 
-			suite.Assert().Equal(tc.expSkip, skip, "SendGovMsg result skip bool")
-			suite.Assert().Equal(tc.expOpMsgRoute, opMsg.Route, "SendGovMsg result op msg route")
-			suite.Assert().Equal(tc.expOpMsgName, opMsg.Name, "SendGovMsg result op msg name")
-			suite.Assert().Equal(tc.expOpMsgComment, opMsg.Comment, "SendGovMsg result op msg comment")
+			s.Assert().Equal(tc.expSkip, skip, "SendGovMsg result skip bool")
+			s.Assert().Equal(tc.expOpMsgRoute, opMsg.Route, "SendGovMsg result op msg route")
+			s.Assert().Equal(tc.expOpMsgName, opMsg.Name, "SendGovMsg result op msg name")
+			s.Assert().Equal(tc.expOpMsgComment, opMsg.Comment, "SendGovMsg result op msg comment")
 			if !tc.expSkip && !skip {
 				// If we don't expect a skip, and we didn't get one,
 				// get the last gov prop and make sure it's the one we just sent.
 				expMsgs := []sdk.Msg{tc.msg}
-				prop := suite.getLastGovProp()
-				if suite.Assert().NotNil(prop, "last gov prop") {
+				prop := s.getLastGovProp()
+				if s.Assert().NotNil(prop, "last gov prop") {
 					msgs, err := prop.GetMsgs()
-					if suite.Assert().NoError(err, "error from prop.GetMsgs() on the last gov prop") {
-						suite.Assert().Equal(expMsgs[0], msgs[0], "messages in the last gov prop")
+					if s.Assert().NoError(err, "error from prop.GetMsgs() on the last gov prop") {
+						s.Assert().Equal(expMsgs[0], msgs[0], "messages in the last gov prop")
 					}
 				}
 			}
@@ -338,21 +351,95 @@ func (suite *SimTestSuite) TestSimulateMsgAddMarkerProposal() {
 	}
 }
 
+func (s *SimTestSuite) TestSSimulateMsgSetAccountData() {
+	// setup 3 accounts
+	src := rand.NewSource(1)
+	r := rand.New(src)
+	accounts := s.getTestingAccounts(r, 3)
+
+	// Add a marker with deposit permissions so that it can be found by the sim.
+	newMarker := &types.MsgAddFinalizeActivateMarkerRequest{
+		Amount:      sdk.NewInt64Coin("simcoin", 1000),
+		Manager:     accounts[1].Address.String(),
+		FromAddress: accounts[1].Address.String(),
+		MarkerType:  types.MarkerType_RestrictedCoin,
+		AccessList: []types.AccessGrant{
+			{
+				Address: accounts[1].Address.String(),
+				Permissions: types.AccessList{
+					types.Access_Mint, types.Access_Burn, types.Access_Deposit, types.Access_Withdraw,
+					types.Access_Delete, types.Access_Admin, types.Access_Transfer,
+				},
+			},
+		},
+		SupplyFixed:            true,
+		AllowGovernanceControl: true,
+		AllowForcedTransfer:    false,
+		RequiredAttributes:     nil,
+	}
+	markerMsgServer := keeper.NewMsgServerImpl(s.app.MarkerKeeper)
+	_, err := markerMsgServer.AddFinalizeActivateMarker(s.ctx, newMarker)
+	s.Require().NoError(err, "AddFinalizeActivateMarker")
+
+	// begin a new block
+	s.app.BeginBlock(abci.RequestBeginBlock{Header: tmproto.Header{Height: s.app.LastBlockHeight() + 1, AppHash: s.app.LastCommitID().Hash}})
+
+	args := s.getWeightedOpsArgs()
+	// execute operation
+	op := simulation.SimulateMsgSetAccountData(s.app.MarkerKeeper, &args)
+	operationMsg, futureOperations, err := op(r, s.app.BaseApp, s.ctx, accounts, "")
+	s.Require().NoError(err, "SimulateMsgSetAccountData op(...) error")
+	s.LogOperationMsg(operationMsg)
+
+	var msg types.MsgSetAccountDataRequest
+	s.Require().NoError(s.app.AppCodec().UnmarshalJSON(operationMsg.Msg, &msg), "UnmarshalJSON(operationMsg.Msg)")
+
+	s.Assert().True(operationMsg.OK, "operationMsg.OK")
+	s.Assert().Equal(sdk.MsgTypeURL(&msg), operationMsg.Name, "operationMsg.Name")
+	s.Assert().Equal("simcoin", msg.Denom, "msg.Denom")
+	s.Assert().Equal("", msg.Value, "msg.Value")
+	s.Assert().Equal(accounts[1].Address.String(), msg.Signer, "msg.Signer")
+	s.Assert().Equal(sdk.MsgTypeURL(&msg), operationMsg.Route, "operationMsg.Route")
+	s.Assert().Len(futureOperations, 0, "futureOperations")
+}
+
+func (s *SimTestSuite) getTestingAccounts(r *rand.Rand, n int) []simtypes.Account {
+	accounts := simtypes.RandomAccounts(r, n)
+
+	initAmt := sdk.TokensFromConsensusPower(1000000, sdk.DefaultPowerReduction)
+	initCoins := sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, initAmt))
+
+	// add coins to the accounts
+	for _, account := range accounts {
+		acc := s.app.AccountKeeper.NewAccountWithAddress(s.ctx, account.Address)
+		s.app.AccountKeeper.SetAccount(s.ctx, acc)
+		err := testutil.FundAccount(s.app.BankKeeper, s.ctx, account.Address, initCoins)
+		s.Require().NoError(err)
+	}
+
+	return accounts
+}
+
+func TestSimTestSuite(t *testing.T) {
+	suite.Run(t, new(SimTestSuite))
+}
+
 // getWeightedOpsArgs creates a standard WeightedOpsArgs.
-func (suite *SimTestSuite) getWeightedOpsArgs() simulation.WeightedOpsArgs {
+func (s *SimTestSuite) getWeightedOpsArgs() simulation.WeightedOpsArgs {
 	return simulation.WeightedOpsArgs{
 		AppParams:  make(simtypes.AppParams),
-		JSONCodec:  suite.app.AppCodec(),
-		ProtoCodec: codec.NewProtoCodec(suite.app.InterfaceRegistry()),
-		AK:         suite.app.AccountKeeper,
-		BK:         suite.app.BankKeeper,
-		GK:         suite.app.GovKeeper,
+		JSONCodec:  s.app.AppCodec(),
+		ProtoCodec: codec.NewProtoCodec(s.app.InterfaceRegistry()),
+		AK:         s.app.AccountKeeper,
+		BK:         s.app.BankKeeper,
+		GK:         s.app.GovKeeper,
+		AttrK:      s.app.AttributeKeeper,
 	}
 }
 
 // getLastGovProp gets the last gov prop to be submitted.
-func (suite *SimTestSuite) getLastGovProp() *govtypes.Proposal {
-	props := suite.app.GovKeeper.GetProposals(suite.ctx)
+func (s *SimTestSuite) getLastGovProp() *govtypes.Proposal {
+	props := s.app.GovKeeper.GetProposals(s.ctx)
 	if len(props) == 0 {
 		return nil
 	}
@@ -360,12 +447,12 @@ func (suite *SimTestSuite) getLastGovProp() *govtypes.Proposal {
 }
 
 // freshCtx creates a new context and sets it to this SimTestSuite's ctx field.
-func (suite *SimTestSuite) freshCtx() {
-	suite.ctx = suite.app.BaseApp.NewContext(false, tmproto.Header{})
+func (s *SimTestSuite) freshCtx() {
+	s.ctx = s.app.BaseApp.NewContext(false, tmproto.Header{})
 }
 
 // createTestingAccountsWithPower creates new accounts with the specified power (coins amount).
-func (suite *SimTestSuite) createTestingAccountsWithPower(r *rand.Rand, count int, power int64) []simtypes.Account {
+func (s *SimTestSuite) createTestingAccountsWithPower(r *rand.Rand, count int, power int64) []simtypes.Account {
 	accounts := simtypes.RandomAccounts(r, count)
 
 	initAmt := sdk.TokensFromConsensusPower(power, sdk.DefaultPowerReduction)
@@ -373,9 +460,9 @@ func (suite *SimTestSuite) createTestingAccountsWithPower(r *rand.Rand, count in
 
 	// add coins to the accounts
 	for _, account := range accounts {
-		acc := suite.app.AccountKeeper.NewAccountWithAddress(suite.ctx, account.Address)
-		suite.app.AccountKeeper.SetAccount(suite.ctx, acc)
-		suite.Require().NoError(testutil.FundAccount(suite.app.BankKeeper, suite.ctx, account.Address, initCoins))
+		acc := s.app.AccountKeeper.NewAccountWithAddress(s.ctx, account.Address)
+		s.app.AccountKeeper.SetAccount(s.ctx, acc)
+		s.Require().NoError(testutil.FundAccount(s.app.BankKeeper, s.ctx, account.Address, initCoins))
 	}
 
 	return accounts